pyproject.toml
[project]
name = "paperless-annotations"
version = "0.1.0"
description = "A Django app to add annotation capabilities to Paperless-ngx."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-ez-tasks",
    "django-ninja>=1.5.1",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "namedatomiclock>=1.1.3",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]

[tool.uv.sources]
django-ez-tasks = { git = "https://github.com/al-eax/django_ez_tasks.git" }
